make percent uncertainties always positive (also for negative variables)

+++ punpy/digital_effects_table/tests/variables.py
@@ -256,7 +256,7 @@ L0_RAD_VARIABLES = {
     },
     "digital_number": {
         "dim": [WL_DIM, SCAN_DIM],
-        "dtype": np.uint32,
+        "dtype": np.float32,
         "attributes": {
             "standard_name": "digital_number",
             "long_name": "Digital number, raw data",
